refactor(sap-aicore-embedder): extract orchestration backend into its own module
